streetcred_backend/backend/streetcred/pyproject.toml
[project]
name = "streetcred"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.7",
    "django-ninja>=1.4.3",
    "dotenv>=0.9.9",
    "folium>=0.20.0",
    "google-cloud-aiplatform>=1.38.0",
    "google-generativeai>=0.8.3",
    "langchain>=0.3.27",
    "langchain-community>=0.3.30",
    "langchain-google-genai>=2.0.5,<2.1.0",
    "langchain-openai>=0.3.34",
    "pillow>=11.3.0",
    "pygeohash>=3.2.0",
    "python-dotenv>=1.1.1",
    "python-geohash>=0.8.5",
    "supabase>=2.21.1",
]
